chore: rename file to avoid SonarCloud getting confused with identically named file in other projects, causing test coverage to drop to 0%
