+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/__tests__/shared/resources/queue/appTaskQueueSagas.test.ts
@@ -5,7 +5,7 @@ import { startInitialAppTaskQueueFulfilled } from '../../../../src/shared/resour
 
 import TextResourcesActions from '../../../../src/shared/resources/textResources/textResourcesActions';
 import ProfileActions from '../../../../src/shared/resources/profile/profileActions';
-import { profileApiUrl } from '../../../../src/utils/urlHelper';
+import { profileApiUrl } from '../../../../src/utils/appUrlHelper';
 import LanguageActions from '../../../../src/shared/resources/language/languageActions';
 import ApplicationMetadataActionDispatcher from '../../../../src/shared/resources/applicationMetadata/actions';
 import PartyActions from '../../../../src/shared/resources/party/partyActions';
