+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/features/entrypoint/Entrypoint.tsx
@@ -20,7 +20,7 @@ import {
 import {
   getActiveInstancesUrl,
   getPartyValidationUrl,
-} from 'src/utils/urlHelper';
+} from 'src/utils/appUrlHelper';
 import { Form } from '../form/containers/Form';
 import { updateValidations } from '../form/validation/validationSlice';
 import Instantiate from '../instantiate/containers';
