+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/features/form/rules/fetch/fetchRulesSagas.ts
@@ -2,7 +2,7 @@ import { SagaIterator } from 'redux-saga';
 import { call, takeLatest, select, put } from 'redux-saga/effects';
 import { IInstance } from 'altinn-shared/types';
 import { IApplicationMetadata } from 'src/shared/resources/applicationMetadata';
-import { getRulehandlerUrl } from 'src/utils/urlHelper';
+import { getRulehandlerUrl } from 'src/utils/appUrlHelper';
 import { get } from '../../../../utils/networking';
 import { getRuleModelFields } from '../../../../utils/rules';
 import Actions from '../rulesActions';
