+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/features/form/validation/singleField/singleFieldValidationSagas.ts
@@ -2,7 +2,7 @@ import { SagaIterator } from 'redux-saga';
 import { call, put, select, takeLatest } from 'redux-saga/effects';
 import { AxiosRequestConfig } from 'axios';
 import { IRuntimeState, IValidationIssue } from 'src/types';
-import { getDataValidationUrl } from 'src/utils/urlHelper';
+import { getDataValidationUrl } from 'src/utils/appUrlHelper';
 import { get } from 'src/utils/networking';
 import { mapDataElementValidationToRedux, mergeValidationObjects } from '../../../../utils/validation';
 import { runSingleFieldValidation,
