+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/shared/resources/queue/appTask/appTaskQueueSagas.ts
@@ -8,7 +8,7 @@ import TextResourcesActions from '../../textResources/textResourcesActions';
 import ProfileActions from '../../profile/profileActions';
 import LanguageActions from '../../language/languageActions';
 import PartyActions from '../../party/partyActions';
-import { profileApiUrl } from '../../../../utils/urlHelper';
+import { profileApiUrl } from '../../../../utils/appUrlHelper';
 
 export function* startInitialAppTaskQueueSaga(): SagaIterator {
   yield put(ApplicationSettingsActions.fetchApplicationSettings());
