+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/utils/appIrlHelper.test.ts
@@ -6,7 +6,7 @@ import {
   getFetchFormDynamicsUrl,
   getLayoutSettingsUrl,
   getHostname,
-} from './urlHelper';
+} from './appUrlHelper';
 
 describe('Frontend urlHelper.ts', () => {
   describe('getEnvironmentLoginUrl', () => {
