+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/utils/appUrlHelper.ts
